consolidate to_numpy
