+++ pandas/core/arrays/boolean.py
@@ -18,9 +18,7 @@ from pandas.core.dtypes.common import (
     is_integer_dtype,
     is_list_like,
     is_numeric_dtype,
-    is_object_dtype,
     is_scalar,
-    is_string_dtype,
     pandas_dtype,
 )
 from pandas.core.dtypes.dtypes import register_extension_dtype
@@ -296,91 +294,6 @@ class BooleanArray(BaseMaskedArray):
     @classmethod
     def _from_factorized(cls, values, original: "BooleanArray"):
         return cls._from_sequence(values, dtype=original.dtype)
-
-    def _formatter(self, boxed=False):
-        return str
-
-    def to_numpy(
-        self, dtype=None, copy=False, na_value: "Scalar" = lib.no_default,
-    ):
-        """
-        Convert to a NumPy Array.
-
-        By default converts to an object-dtype NumPy array. Specify the `dtype` and
-        `na_value` keywords to customize the conversion.
-
-        Parameters
-        ----------
-        dtype : dtype, default object
-            The numpy dtype to convert to.
-        copy : bool, default False
-            Whether to ensure that the returned value is a not a view on
-            the array. Note that ``copy=False`` does not *ensure* that
-            ``to_numpy()`` is no-copy. Rather, ``copy=True`` ensure that
-            a copy is made, even if not strictly necessary. This is typically
-            only possible when no missing values are present and `dtype`
-            is a boolean dtype.
-        na_value : scalar, optional
-             Scalar missing value indicator to use in numpy array. Defaults
-             to the native missing value indicator of this array (pd.NA).
-
-        Returns
-        -------
-        numpy.ndarray
-
-        Examples
-        --------
-        An object-dtype is the default result
-
-        >>> a = pd.array([True, False], dtype="boolean")
-        >>> a.to_numpy()
-        array([True, False], dtype=object)
-
-        When no missing values are present, a boolean dtype can be used.
-
-        >>> a.to_numpy(dtype="bool")
-        array([ True, False])
-
-        However, requesting a bool dtype will raise a ValueError if
-        missing values are present and the default missing value :attr:`NA`
-        is used.
-
-        >>> a = pd.array([True, False, pd.NA], dtype="boolean")
-        >>> a
-        <BooleanArray>
-        [True, False, NA]
-        Length: 3, dtype: boolean
-
-        >>> a.to_numpy(dtype="bool")
-        Traceback (most recent call last):
-        ...
-        ValueError: cannot convert to bool numpy array in presence of missing values
-
-        Specify a valid `na_value` instead
-
-        >>> a.to_numpy(dtype="bool", na_value=False)
-        array([ True, False, False])
-        """
-        if na_value is lib.no_default:
-            na_value = libmissing.NA
-        if dtype is None:
-            dtype = object
-        if self._hasna:
-            if (
-                not (is_object_dtype(dtype) or is_string_dtype(dtype))
-                and na_value is libmissing.NA
-            ):
-                raise ValueError(
-                    f"cannot convert to '{dtype}'-dtype NumPy array "
-                    "with missing values. Specify an appropriate 'na_value' "
-                    "for this dtype."
-                )
-            # don't pass copy to astype -> always need a copy since we are mutating
-            data = self._data.astype(dtype)
-            data[self._mask] = na_value
-        else:
-            data = self._data.astype(dtype, copy=copy)
-        return data
 
     _HANDLED_TYPES = (np.ndarray, numbers.Number, bool, np.bool_)
